devtools: code cleanup
