+++ tmtc-c2a/devtools_frontend/crates/wasm-opslang/src/util.rs
@@ -1,0 +1,6 @@
+#[allow(unused_macros)]
+macro_rules! log {
+    ( $( $t:tt )* ) => {
+        web_sys::console::log_1(&format!( $( $t )* ).into());
+    }
+}
